requirements.txt
requests>=2.25.0
tabulate>=0.8.0
questionary>=2.0.0
python-dotenv>=1.0.0
openpyxl>=3.0.0
dicttoxml>=1.7.0
openai>=1.0.0
